Summarized the MoU
This is my proposal for a shorter and more simple version of the MoU. I have radically cut and reworded many things, so I understand it may not feel as elaborated as the original version; however I tried to include every aspect in the document.  Comments and changes are, following our principles of collaboration, obviously very welcome. Keep on rocking, funkies!